refactor: 추천 관련 API 폴더 이동
